Initial commit for Jakarta Security 4.0 - Multiple HttpAuthenticationMechanisms.
All functionality works for Jakarta Security 1/2/3.0 and 4.0.

A new service called HttpAuthenticationMechanismHandlerService was
inserted into AuthModule for the interface methods validateRequest,
cleanSubject and secureResponse.

The HAMHandlerService is fetched from the OSGi, and differs for Jakarta
Security 1/2/3.0 and 4.0.

For Jakarta Security 1/2/3.0, the HAMHandlerService implements the
identical functionality as previous for the methods methods above.

For Jakarta Security 4.0, the HAMHandlerService finds the implementation
( in-built or application specified ) of the
HttpAuthenticationmechansimHandler interface ( this is Jakarta 4.0
specified ) from the CDI, and defers to this implementation to execute
the methods above.

For Jakarta Security 4.0, an additional completely internal HAM handler
is created also with the @Alternative annotation to ensure the check for
multiple HAM Handlers is skipped ( as per the specification ).